+++ lib/features/restauran_tour/view/restauran_tour.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:restaurantour/features/restauran_tour/screens/all_restaurants_screen.dart';
-import 'package:restaurantour/features/restauran_tour/screens/my_favorites_screen.dart';
+import 'package:restaurantour/features/restauran_tour/view/all_restaurants_screen.dart';
+import 'package:restaurantour/features/restauran_tour/view/my_favorites_screen.dart';
 import 'package:restaurantour/theme/app_color.dart';
 
 class RestauranTour extends StatelessWidget {
