+++ test/widget_test.dart
@@ -6,13 +6,12 @@
 // tree, read text, and verify that the values of widget properties are correct.
 
 import 'package:flutter_test/flutter_test.dart';
-
-import 'package:restaurantour/main.dart';
+import 'package:restaurantour/features/restauran_tour/view/restauran_tour.dart';
 
 void main() {
   testWidgets('Page loads', (WidgetTester tester) async {
     // Build our app and trigger a frame.
-    await tester.pumpWidget(const Restaurantour());
+    await tester.pumpWidget(const RestauranTour());
 
     // Verify that tests will run
     expect(find.text('Fetch Restaurants'), findsOneWidget);
